add: 增加visionpro的支持
